+++ src/navigation/types.ts
@@ -4,6 +4,6 @@ export type RootStackParamList = {
   Login: undefined;
   Signup: undefined;
   Home: undefined; // 로그인 성공 시 이동할 화면이 있다면 추가
-  JoinTeam: undefined;
+  JoinTeam:undefined;
 };
