+++ src/screens/JoinTeamScreen.tsx
@@ -5,6 +5,8 @@ import { View, Text, TextInput, TouchableOpacity, Alert, StyleSheet } from 'reac
 import { NativeStackNavigationProp } from '@react-navigation/native-stack';
 import { RootStackParamList } from '../navigation/types';
 import { joinTeamByCode } from '../api/teamApi';
+import styles from '../styles/JoinTeamScreem.styles';
+
 
 type JoinTeamScreenProps = {
   navigation: NativeStackNavigationProp<RootStackParamList, 'JoinTeam'>;
@@ -22,7 +24,7 @@ export default function JoinTeamScreen({ navigation }: JoinTeamScreenProps) {
     try {
       await joinTeamByCode(code);
       Alert.alert('팀 가입 완료', '성공적으로 팀에 가입했습니다!');
-      navigation.navigate('Home'); // 필요 시 다른 화면으로 이동
+      navigation.navigate('Home',); // 필요 시 다른 화면으로 이동
     } catch (error: any) {
       Alert.alert('가입 실패', error.message || '팀 가입 중 오류가 발생했습니다.');
     }
@@ -43,36 +45,3 @@ export default function JoinTeamScreen({ navigation }: JoinTeamScreenProps) {
     </View>
   );
 }
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    padding: 30,
-    justifyContent: 'center',
-    backgroundColor: '#fffbe0',
-  },
-  title: {
-    fontSize: 22,
-    marginBottom: 24,
-    textAlign: 'center',
-  },
-  input: {
-    borderWidth: 1,
-    borderColor: '#999',
-    padding: 12,
-    borderRadius: 8,
-    backgroundColor: '#fff',
-    marginBottom: 20,
-  },
-  button: {
-    backgroundColor: '#f4a261',
-    padding: 14,
-    borderRadius: 10,
-    alignItems: 'center',
-  },
-  buttonText: {
-    color: '#fff',
-    fontSize: 16,
-    fontWeight: 'bold',
-  },
-});
