+++ src/screens/SignupScreen.tsx
@@ -1,11 +1,11 @@
 import React, { useState } from 'react';    
 import {View,TextInput,Alert,Text,TouchableOpacity,StyleSheet,} from 'react-native';
 import { RadioButton } from 'react-native-paper';
-import DropDownPicker, { ItemType } from 'react-native-dropdown-picker';
+import { ItemType } from 'react-native-dropdown-picker';
 import { NativeStackNavigationProp } from '@react-navigation/native-stack';
 import { RootStackParamList } from '../navigation/types';
 import { signup } from '../api/authApi';  // api 함수 import
-
+import styles from '../styles/SignupScreen.styles'
 
 type SignupScreenProps = {
   navigation: NativeStackNavigationProp<RootStackParamList, 'Signup'>;
@@ -51,7 +51,7 @@ export default function SignupScreen({ navigation }: SignupScreenProps) {
   if (!y || !m || !d || isNaN(year) || isNaN(month) || isNaN(day) || year < 1900 || year > new Date().getFullYear() || month < 1 || month > 12) {
     isValid = false;
   } else {
-    const date = new Date(year, month - 1, day);
+    const date = new Date(year, month - 1, day); //다시 공부 
     if (date.getFullYear() !== year || date.getMonth() + 1 !== month || date.getDate() !== day) {
       isValid = false;
     }
@@ -86,12 +86,11 @@ const validateEmailId = (id: string) => {
       Alert.alert('입력 오류', '모든 항목을 입력하세요.');
       return;
     } 
-
     //생년월일 오류(중복적이지 않나?)
-    if (birthError !== '') {
-      Alert.alert('입력 오류', '올바른 생년월일을 입력하세요.');
-      return;
-    }
+    // if (birthError !== '') {
+    //   Alert.alert('입력 오류', '올바른 생년월일을 입력하세요.');
+    //   return;
+    // }
     //비번 오류
     const passwordErr = getPasswordErrorMessage(password);
       if (passwordErr !== '') {
@@ -117,6 +116,7 @@ const validateEmailId = (id: string) => {
   };
 
   const onlyNumber = (text: string) => text.replace(/[^0-9]/g, ''); //선택사항 나중에 지우기
+  
   return (
     
     <View style={styles.container}>
@@ -270,92 +270,3 @@ const validateEmailId = (id: string) => {
   );
 
 }
-
-const styles = StyleSheet.create({
-  container: { padding: 50, flex: 1, backgroundColor: '#fff',justifyContent: 'center', },
-  title: { fontSize: 24, marginBottom: 20, textAlign: 'center' },
-  label: {
-    marginBottom: 6,
-    fontWeight: '600',
-  },
-  input: {
-    height: 48,
-    borderBottomWidth: 1,
-    borderColor: '#aaa',
-    marginBottom: 16,
-    paddingHorizontal: 8,
-    fontSize: 16,
-  },
-  radioGroup: {
-    flexDirection: 'column',
-    justifyContent: 'space-around',
-    
-  },
-  radioOption: {
-    flexDirection: 'row',
-    alignItems: 'center',
-  },
-  genderBox: {
-    backgroundColor: 'rgba(255, 255, 255, 0.7)',
-    borderRadius: 10,
-    padding: 1,
-    marginBottom: 20,
-    borderWidth: 1,
-    borderColor: '#ccc',
-  },
-  roundButton: {
-    backgroundColor: '#FF9898',
-    paddingVertical: 12,
-    paddingHorizontal: 30,
-    borderRadius: 30,
-    alignItems: 'center',
-    marginBottom: 10,
-  },
-  roundButtonText: {
-    color: 'white',
-    fontSize: 16,
-    fontWeight: 'bold',
-    
-  },
-  emailRow: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    marginBottom: 20,
-    zIndex: 1000,
-  },
-  emailInput: {
-
-  height: 50,
-  borderWidth: 1,
-  borderColor: '#ccc',
-  paddingHorizontal: 10, // 좌우 패딩만 줌, 위아래 패딩은 0으로
-  paddingVertical: 0,
-  borderTopLeftRadius: 15,
-  borderBottomLeftRadius: 15,
-  borderRightWidth: 0,
-  backgroundColor: 'rgba(255, 255, 255, 0.7)',
-  },
-  dropdownWrapper: {
-  flex: 6,
-  height: 50,
-  zIndex: 1000,
-  },
-  dropdown: {
-  height: 50,
-  borderWidth: 1,
-  borderColor: '#ccc',
-  paddingHorizontal: 10,  // emailInput과 맞춤
-  paddingVertical: 0,
-  borderTopRightRadius: 15,
-  borderBottomRightRadius: 15,
-  borderTopLeftRadius: 0,
-  borderBottomLeftRadius: 0,
-  borderLeftWidth: 0,
-  backgroundColor: 'rgba(255, 255, 255, 0.7)',
-},
-  dropdownContainer: {
-    borderWidth: 1,
-    borderColor: '#ccc',
-    zIndex: 1000,
-  },
-});
