+++ src/styles/JoinTeamScreem.styles.ts
@@ -1,0 +1,43 @@
+import { StyleSheet } from 'react-native';
+
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    paddingTop: 160,
+    paddingHorizontal: 24,
+    backgroundColor: '#fff',
+  },
+  title: {
+   fontSize: 22,
+    fontWeight: 'bold',
+    textAlign: 'center',
+    marginBottom: 12,
+    color: '#111',
+  },
+  input: {
+    backgroundColor: '#f7f8fa',
+    borderRadius: 16,
+    height: 52,
+    paddingHorizontal: 16,
+    fontSize: 16,
+    color: '#333',
+    marginTop :100
+  },
+  button: {
+    backgroundColor: '#FF9898',
+    paddingVertical: 14,
+    paddingHorizontal: 140,
+    borderRadius: 30,
+    alignItems: 'center',
+    marginTop: 350,
+  },
+  buttonText: {
+    color: '#fff',
+    textAlign: 'center',
+    fontSize: 16,
+    fontWeight: 'bold',
+  },
+});
+
+export default styles;
